% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_style_json.R
\name{extract_style_json}
\alias{extract_style_json}
\title{Extract JSON from snazzyMaps}
\usage{
extract_style_json(mapRef)
}
\arguments{
\item{mapRef}{either the full Snazzymap page URL, or the ID of the map layer}
}
\description{
For a given web page, this function scrapes the JSON files,
  carries out minor reformatting, and convert the object into
  a list.
}
\author{
Michael Harper
}
