% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snazzymap.R
\name{get_snazzymap}
\alias{get_snazzymap}
\title{Create a stylised google basemap}
\usage{
get_snazzymap(..., mapRef = NULL)
}
\arguments{
\item{...}{arguments to be supplied to get_googlemap}

\item{mapRef}{a URL or map ID for the snazzymap style}
}
\description{
A wrapper for the \@code{get_googlemap}
}
\seealso{
get_googlemap
}
\author{
Michael Harper
}
