% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_style_sm.R
\name{ggmap_style_sm}
\alias{ggmap_style_sm}
\title{Create a google maps style string}
\usage{
ggmap_style_sm(mapRef)
}
\arguments{
\item{mapRef}{the full Snazzymap page URL, or the ID of the map layer}
}
\description{
Provided with a SnazzyMap ID, this function will return a formatted stylestring
 that can be used within the \code{get_googlemap} function.
}
\seealso{
\code{\link[ggmap]{get_googlemap}}
}
