% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_styles.R
\name{list_styles}
\alias{list_styles}
\title{List all the ggmap styles available}
\usage{
list_styles(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Shows the names, ID and description of the cached ggmap styles.
}
