% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_url.R
\name{smart_url}
\alias{smart_url}
\title{Constructs the full SnazzyMaps url if only provided with ID}
\usage{
smart_url(mapRef)
}
\arguments{
\item{mapRef}{the numeric ID or full URL of the Snazzy Map style}
}
\description{
Creates the full ID of the web page, being provided either the full 
  URL or only the page ID. The resulting page is validated to ensure
  that the page exists.
}
