% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_style_sm.R
\name{ggmap_style_sm}
\alias{ggmap_style_sm}
\title{Create a style string for}
\usage{
ggmap_style_sm(mapRef)
}
\arguments{
\item{mapRef}{the full Snazzymap page URL, or the ID of the map layer}
}
\description{
Create a style string for
}
