% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_style_string.R
\name{ggmap_style_string}
\alias{ggmap_style_string}
\title{Create style from JSON}
\usage{
ggmap_style_string(style_list)
}
\arguments{
\item{style_list}{A formatted JSON object within R, as loaded by the
extract_style_json function.}
}
\description{
Creates a google maps style URL string from the formatted JSON table. 
  The output of this function can be used within the style argument of
  get_googlemap.
}
